♻️ refactor [#10] 사이드바 컴포넌트 import 링크 수정
